+++ packages/url_launcher/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.1.0
+
+* Added `universalLinksOnly` setting.
+* Updated `launch` to return `Future<bool>`.
+
 ## 4.0.3
 
 * Fixed launch url fail for Android: `launch` now assert activity not null and using activity to startActivity.